Move bleve-related commands under a common subcommand.
